requirements.txt
Django==2.2.7
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
pkg-resources==0.0.0
pytesseract==0.3.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
tesseract==0.1.3
